Remove extension
